requirements.txt
pytest
git+https://github.com/dlsyscourse/mugrade.git#egg=mugrade
pybind11
numdifftools
requests
numpy
pandas
python-dotenv
torch
torchrec
awswrangler
pyyaml
tabulate